fixed Partners->Causes, Profile->Cause
