+++ client/src/components/Nav/nav.js
@@ -7,7 +7,7 @@ import Auth from '../../utils/auth';
 const LoggedInNav = () => (
     <>
         <NavLink to='/' className="links" >Home</NavLink>
-        <NavLink to='/partners' className="links" >Partners</NavLink>
+        <NavLink to='/causes' className="links" >Partners</NavLink>
         <NavLink to='/my-cause' className="links">My Cause</NavLink>
         <button className='btn btn-light' onClick={() => Auth.logout()}>Logout</button>
         <h3>{Auth.getProfile().email}</h3>
@@ -17,7 +17,7 @@ const LoggedInNav = () => (
 const LoggedOutNav = () => (
     <>
         <NavLink to='/' className="links" >Home</NavLink>
-        <NavLink to='/partners' className="links" >Partners</NavLink>
+        <NavLink to='/causes' className="links" >Partners</NavLink>
         <NavLink to='/login' className="links">Login</NavLink>
         <NavLink to='/signup' className="links">Sign Up</NavLink>
     </>
